Remove redundant interfaces
Remove Mapper and MappingDocument interfaces and refactor their implementation as the go to classes.
