+++ src/test/java/r2graph/util/MappingDocumentTest.java
@@ -5,8 +5,6 @@ import org.apache.jena.rdf.model.ModelFactory;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.io.InputStream;
-
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
@@ -14,16 +12,16 @@ import static org.hamcrest.MatcherAssert.assertThat;
  * Tests for MappingDocument code. Currently skeletal till further
  * iteration of the project or feature.
  */
-public abstract class AbstractMappingDocumentTest {
+public class MappingDocumentTest {
 
+    private final String r2rmlFilename = "../../r2rml_input_tableName_singlePom_01.ttl";
     private MappingDocument mappingDocument;
-
-    public abstract MappingDocument createInstance();
 
     @Before
     public void setUp() throws Exception {
-        mappingDocument = createInstance();
+        mappingDocument = new MappingDocument(getClass().getResourceAsStream(r2rmlFilename));
     }
+
 
     /**
      * Tests that the MappingDocument returns a populated {@code Model} containing
@@ -32,8 +30,8 @@ public abstract class AbstractMappingDocumentTest {
      */
     @Test
     public void WhenMappingDocumentExists_ShouldReturnGraph() {
-        InputStream in = getClass().getResourceAsStream("../../r2rml_input_tableName_singlePom_01.ttl");
-        Model expected = ModelFactory.createDefaultModel().read(in, "http://example.org/ns#", "TTL");
+        Model expected = ModelFactory.createDefaultModel().read(
+                getClass().getResourceAsStream(r2rmlFilename), null, "TTL");
 
         Model result = mappingDocument.getMappingGraph();
